test: multi contract keystore
